Updating PR relative to comments, not yet addressed PBS
